+++ src/api/index.ts
@@ -20,3 +20,4 @@ export * from './mentions';
 export * from './search';
 export * from './subscription';
 export * from './publicProfile';
+export * from './shortLink';