+++ src/api/routes.ts
@@ -25,6 +25,8 @@ export const CATEGORIES_ROUTE = `${ITEMS_ROUTE}/categories`;
 export const ETHERPAD_ROUTE = `${ITEMS_ROUTE}/etherpad`;
 export const COLLECTIONS_ROUTE = `collections`;
 export const buildAppListRoute = `${APPS_ROUTE}/list`;
+export const SHORT_LINKS_ROUTE = `${ITEMS_ROUTE}/short-links`;
+export const SHORT_LINKS_LIST_ROUTE = `${SHORT_LINKS_ROUTE}/list`;
 
 export const buildPostItemRoute = (parentId?: UUID) => {
   let url = ITEMS_ROUTE;
@@ -400,6 +402,15 @@ export const PUBLIC_PROFILE_ROUTE = `profile`;
 export const GET_OWN_PROFILE = `${PUBLIC_PROFILE_ROUTE}/own`;
 export const buildGetPublicProfileRoute = (memberId: UUID) =>
   `${PUBLIC_PROFILE_ROUTE}/${memberId}`;
+export const buildGetShortLinkAvailableRoute = (alias: string) =>
+  `${SHORT_LINKS_ROUTE}/available/${alias}`;
+export const buildGetShortLinksItemRoute = (itemId: string) =>
+  `${SHORT_LINKS_LIST_ROUTE}/${itemId}`;
+export const buildDeleteShortLinkRoute = (alias: string) =>
+  `${SHORT_LINKS_ROUTE}/${alias}`;
+export const buildPostShortLinkRoute = () => `${SHORT_LINKS_ROUTE}`;
+export const buildPatchShortLinkRoute = (alias: string) =>
+  `${SHORT_LINKS_ROUTE}/${alias}`;
 
 export const API_ROUTES = {
   APPS_ROUTE,
@@ -417,6 +428,7 @@ export const API_ROUTES = {
   buildDeleteItemsRoute,
   buildDeleteItemTagRoute,
   buildDeleteMemberRoute,
+  buildDeleteShortLinkRoute,
   buildDownloadAvatarRoute,
   buildDownloadFilesRoute,
   buildDownloadItemThumbnailRoute,
@@ -454,6 +466,8 @@ export const API_ROUTES = {
   buildGetMostRecentPublishedItemsRoute,
   buildGetPlanRoute,
   buildGetPublishedItemsForMemberRoute,
+  buildGetShortLinkAvailableRoute,
+  buildGetShortLinksItemRoute,
   buildImportH5PRoute,
   buildImportZipRoute,
   buildItemPublishRoute,
@@ -464,6 +478,7 @@ export const API_ROUTES = {
   buildPatchInvitationRoute,
   buildPatchItemChatMessageRoute,
   buildPatchMember,
+  buildPatchShortLinkRoute,
   buildPostEtherpadRoute,
   buildPostInvitationsRoute,
   buildPostItemAction,
@@ -477,6 +492,7 @@ export const API_ROUTES = {
   buildPostItemTagRoute,
   buildPostItemValidationRoute,
   buildPostManyItemMembershipsRoute,
+  buildPostShortLinkRoute,
   buildPutItemLoginSchemaRoute,
   buildRecycleItemRoute,
   buildRecycleItemsRoute,