+++ src/hooks/index.ts
@@ -19,6 +19,7 @@ import configureMentionsHooks from './mention';
 import configurePlanHooks from './plan';
 import configurePublicProfileHooks from './publicProfile';
 import configureKeywordSearchHooks from './search';
+import configureShortLinkHooks from './shortLink';
 
 export default (
   queryConfig: QueryClientConfig,
@@ -53,5 +54,6 @@ export default (
     ...memberHooks,
     ...configurePlanHooks(queryConfig),
     ...configurePublicProfileHooks(queryConfig),
+    ...configureShortLinkHooks(queryConfig),
   };
 };